docs: properly formatted
